Overhaul Makefiles
We now have our own top-level Makefile instead of the one generated
by the Rack SDK's helper.py script. This now properly builds the Rust
library as a dependency of the C++ portion of the plugin.

We also have a better cross-compilation story in the libpotential
Makefile, so we can build a specific target if that's what we're after,
and otherwise we assume it's a local build and optimize for the native
CPU.