Change Port polyphony methods to safer semantics
This changes the internal Port.channels type to an `Option<NonZeroU8>`,
which is guaranteed to have the same layout as a `u8` but has special
handling for the 0 hole. VCV rack uses a polyphony count of 0 to signal
a disconnected port, and handling that as a special case without a
special type led to some unintentional misuse in PolyShuffle that caused
a runtime crash owing to the zero-length slice the old interface could
return.

We also change most of the methods of InputPort and OutputPort to more
ergonomically deal with this new internal model.